+++ src/shared/java/org/truffleruby/shared/options/OptionsCatalog.java
@@ -75,6 +75,7 @@ public class OptionsCatalog {
     public static final OptionKey<Boolean> CEXTS_LOG_LOAD_KEY = new OptionKey<>(false);
     public static final OptionKey<Boolean> CEXTS_LOG_WARNINGS_KEY = new OptionKey<>(false);
     public static final OptionKey<Boolean> ARGV_GLOBALS_KEY = new OptionKey<>(false);
+    public static final OptionKey<Boolean> GETS_LOOP_KEY = new OptionKey<>(false);
     public static final OptionKey<Boolean> IGNORE_LINES_BEFORE_RUBY_SHEBANG_KEY = new OptionKey<>(false);
     public static final OptionKey<Boolean> SYNTAX_CHECK_KEY = new OptionKey<>(false);
     public static final OptionKey<String[]> ARGV_GLOBAL_VALUES_KEY = new OptionKey<>(new String[]{}, StringArrayOptionType.INSTANCE);
@@ -523,6 +524,13 @@ public class OptionsCatalog {
     public static final OptionDescriptor ARGV_GLOBALS = OptionDescriptor
             .newBuilder(ARGV_GLOBALS_KEY, "ruby.argv-globals")
             .help("Parse options in script argv into global variables (configured by the -s Ruby option)")
+            .category(OptionCategory.INTERNAL)
+            .stability(OptionStability.EXPERIMENTAL)
+            .build();
+
+    public static final OptionDescriptor GETS_LOOP = OptionDescriptor
+            .newBuilder(GETS_LOOP_KEY, "ruby.gets-loop")
+            .help("assume while gets(); ... end loop around your script (configured by the -n Ruby option)")
             .category(OptionCategory.INTERNAL)
             .stability(OptionStability.EXPERIMENTAL)
             .build();
@@ -1101,6 +1109,8 @@ public class OptionsCatalog {
                 return CEXTS_LOG_WARNINGS;
             case "ruby.argv-globals":
                 return ARGV_GLOBALS;
+            case "ruby.gets-loop":
+                return GETS_LOOP;
             case "ruby.ignore-lines-before-ruby-shebang":
                 return IGNORE_LINES_BEFORE_RUBY_SHEBANG;
             case "ruby.syntax-check":
@@ -1295,6 +1305,7 @@ public class OptionsCatalog {
             CEXTS_LOG_LOAD,
             CEXTS_LOG_WARNINGS,
             ARGV_GLOBALS,
+            GETS_LOOP,
             IGNORE_LINES_BEFORE_RUBY_SHEBANG,
             SYNTAX_CHECK,
             ARGV_GLOBAL_VALUES,