+++ CHANGELOG.md
@@ -77,6 +77,7 @@ Compatibility:
 * Implemented `Enumerator::Chain`, `Enumerator#+`, and `Enumerable#chain` (#1859, #1858).
 * Implemented `Thread#backtrace_locations` and `Exception#backtrace_locations` (#1556).
 * Implemented `rb_module_new`, `rb_define_class_id`, `rb_define_module_id`, (#1876, @chrisseaton).
+* Implemented `-n` CLI option (#1532).
 
 Performance: